fix: DRY up styles and tweak snapshots
