+++ packages/article-list/src/article-list.web.js
@@ -112,7 +112,8 @@ class ArticleList extends Component {
           elementId: `${article.id}.${index}`
         }));
 
-    if (data.length === 0) return <ArticleListEmptyState message={emptyStateMessage} />;
+    if (data.length === 0)
+      return <ArticleListEmptyState message={emptyStateMessage} />;
 
     const Contents = ({ clientHasRendered, observed, registerNode }) =>
       data.length === 0 ? (
